TIP-236: Move user model to the component
