+++ features/Context/FixturesContext.php
@@ -26,13 +26,11 @@ use Pim\Bundle\CatalogBundle\Entity\GroupType;
 use Pim\Bundle\CommentBundle\Entity\Comment;
 use Pim\Bundle\CommentBundle\Model\CommentInterface;
 use Pim\Bundle\DataGridBundle\Entity\DatagridView;
-use Pim\Bundle\UserBundle\Entity\User;
 use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Builder\EntityWithValuesBuilderInterface;
 use Pim\Component\Catalog\Model\Association;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\AttributeOptionInterface;
-use Pim\Component\Catalog\Model\FamilyInterface;
 use Pim\Component\Catalog\Model\LocaleInterface;
 use Pim\Component\Catalog\Model\ProductInterface;
 use Pim\Component\Catalog\Model\ProductModelInterface;
@@ -45,6 +43,7 @@ use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModel
 use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleCsvExport;
 use Pim\Component\ReferenceData\Model\ReferenceDataInterface;
 use Pim\Component\User\Model\Role;
+use Pim\Component\User\Model\User;
 
 /**
  * A context for creating entities
@@ -1653,7 +1652,7 @@ class FixturesContext extends BaseFixturesContext
     /**
      * @param string $username
      *
-     * @return User
+     * @return \Pim\Component\User\Model\User
      *
      * @Then /^there should be a "([^"]*)" user$/
      */
