+++ src/Pim/Bundle/DataGridBundle/spec/Extension/Formatter/Property/Version/AuthorPropertySpec.php
@@ -4,8 +4,8 @@ namespace spec\Pim\Bundle\DataGridBundle\Extension\Formatter\Property\Version;
 
 use Oro\Bundle\DataGridBundle\Datasource\ResultRecordInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\UserBundle\Entity\User;
 use Pim\Bundle\UserBundle\Manager\UserManager;
+use Pim\Component\User\Model\User;
 use Prophecy\Argument;
 use Symfony\Component\Translation\TranslatorInterface;
