+++ src/Pim/Bundle/EnrichBundle/spec/Normalizer/VersionNormalizerSpec.php
@@ -5,9 +5,9 @@ namespace spec\Pim\Bundle\EnrichBundle\Normalizer;
 use Akeneo\Component\Localization\Presenter\PresenterInterface;
 use Akeneo\Component\Versioning\Model\Version;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\UserBundle\Entity\User;
 use Pim\Bundle\UserBundle\Manager\UserManager;
 use Pim\Component\Catalog\Localization\Presenter\PresenterRegistryInterface;
+use Pim\Component\User\Model\User;
 use Prophecy\Argument;
 use Symfony\Component\Translation\TranslatorInterface;
