+++ src/Pim/Bundle/VersioningBundle/spec/EventSubscriber/AddUserSubscriberSpec.php
@@ -3,9 +3,9 @@
 namespace spec\Pim\Bundle\VersioningBundle\EventSubscriber;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\UserBundle\Entity\User;
 use Pim\Bundle\VersioningBundle\Event\BuildVersionEvent;
 use Pim\Bundle\VersioningBundle\Event\BuildVersionEvents;
+use Pim\Component\User\Model\User;
 use Prophecy\Argument;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
