+++ src/Pim/Component/User/Model/User.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Pim\Bundle\UserBundle\Entity;
+namespace Pim\Component\User\Model;
 
 use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Pim\Bundle\CatalogBundle\Entity\Channel;
+use Pim\Bundle\UserBundle\Entity\Group;
+use Pim\Bundle\UserBundle\Entity\UserInterface;
 use Pim\Component\Catalog\Model\CategoryInterface;
 use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\Catalog\Model\LocaleInterface;
-use Pim\Component\User\Model\GroupInterface;
-use Pim\Component\User\Model\Role;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 /**
